test(compat): move zod-compat tests to zodCompat.test.ts; tighten normalizeRawShapeSchema guard to require ~standard.jsonSchema
